pyproject.toml
[tool.poetry]
name = "s0nh-kernel-builder"
version = "0.3.1"
description = ""
authors = ["seppzer0"]
packages = [{include = "wrapper"}]

[tool.poetry.dependencies]
python = "^3.10"
conan = "~1"
argparse = "*"
requests = "*"
bandit = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
